Map-Reduce/function-master/requirements.txt
functions-framework==3.*
google-cloud==0.34.0
google-cloud-storage==2.13.0